+++ lib/internal/modules/esm/get_format.js
@@ -55,7 +55,7 @@ function defaultGetFormat(url, context, defaultGetFormat) {
   } else if (parsed.protocol === 'file:') {
     const ext = extname(parsed.pathname);
     let format;
-    if (ext === '.js' || ext === '') {
+    if (ext === '.js') {
       format = getPackageType(parsed.href) === TYPE_MODULE ?
         'module' : 'commonjs';
     } else {
